Let client retry when nigori root entity is not returned for NEW_CLIENT GetUpdates.
Due to eventually read consistency, it is possible when we read updates from
dynamoDB, those server created unique entities including nigori root folder
are not returned. Nigori root folder is essential for client to properly
initializing the sync engine with nigori type, so we return error in this case
and let client retry if they cannot get it in their first NEW_CLIENT GetUpdates
request.