Add a redis cache for storing clientID#dataType, mtime pairs
When we commit new entries to a type or when we get updates from dynamoDB, we
will save a (clientID#dataType, mtime) pair into the cache, which represents
the last seen modification time of one specific sync type for a client.
With this information in the cache, we could check if there might be no updates
in the dynamoDB since the client's passed token in the FromProgressMarker when
we handle poll get updates requests. If cached mtime <= client's token, there
might be no updates to query in the DB so we can skip querying dynamoDB to
reduce our RCU cost on the GSI. To be safe, we set ttl for key to be 5 minutes
so we will not skip querying DB forever when things somehow go wrong.